* develop: (67 commits)
  Fix compilation of WebClientBackend
  Fix TalkDeliveryTest
  Correct placeholder gravity
  Raise code version
  Fix placeholder layout
  Allow clearing chat of worldwide group history
  Adapt profile menu for worldwide contacts
  Clear avatar before updating
  Show more types of contact in group management dialog
  Show worldwide acquaintance avatar status in friends list
  Show worldwide acquaintance
  Adapt nearby acquaintance flag to support worldwide acquaintance
  Add database migration for notificationPreference
  Correct worldwide group name in contact selection
  Raise version code
  Show worldwide contacts in contact selection
  Destroy worldwide group according to nearby
  Show worldwide contact as online in contact list
  Check for both nearby and worldwide
  Update circle avatar image
  ...

+++ client/src/main/java/com/hoccer/talk/client/XoClient.java
@@ -1393,6 +1393,8 @@ public class XoClient implements JsonRpcConnection.Listener, TransferListener {
                     TalkGroupPresence groupPresence = groupContact.getGroupPresence();
                     if (groupPresence != null && groupPresence.isTypeNearby()) {
                         destroyNearbyGroup(groupContact);
+                    } else if (groupPresence != null && groupPresence.isTypeWorldwide()) {
+                        destroyWorldwideGroup(groupContact);
                     } else {
                         if (groupPresence != null) {
                             if (groupContact.getGroupMembership().isInvolved() && hasMembersOrMessages(groupContact)) {
@@ -1915,6 +1917,7 @@ public class XoClient implements JsonRpcConnection.Listener, TransferListener {
 
     public void sendEnvironmentUpdate(final TalkEnvironment environment) {
         LOG.debug("sendEnvironmentUpdate()");
+
         if (isReady() && environment != null) {
             if (mEnvironmentUpdateCallPending.compareAndSet(false, true)) {
                 mExecutor.execute(new Runnable() {
@@ -1990,7 +1993,11 @@ public class XoClient implements JsonRpcConnection.Listener, TransferListener {
                 @Override
                 public void run() {
                     try {
-                        mServerRpc.destroyEnvironment(type);
+                        if (TalkEnvironment.TYPE_NEARBY.equals(type)) {
+                            mServerRpc.destroyEnvironment(type);
+                        } else {
+                            mServerRpc.releaseEnvironment(type);
+                        }
                     } catch (Throwable t) {
                         LOG.error("sendDestroyEnvironment: other error", t);
                     }
@@ -2025,7 +2032,7 @@ public class XoClient implements JsonRpcConnection.Listener, TransferListener {
             }
 
             if (groupContact == null) {
-                keepNearbyAcquaintance(clientContact);
+                keepAcquaintance(clientContact);
             }
 
             String messageId = delivery.getMessageId();
@@ -2294,7 +2301,7 @@ public class XoClient implements JsonRpcConnection.Listener, TransferListener {
             messageFailed = false;
 
             if (groupContact == null) {
-                keepNearbyAcquaintance(senderContact);
+                keepAcquaintance(senderContact);
             }
 
         } catch (GeneralSecurityException e) {
@@ -2349,14 +2356,17 @@ public class XoClient implements JsonRpcConnection.Listener, TransferListener {
         sendDownloadAttachmentDeliveryConfirmation(delivery);
     }
 
-    private void keepNearbyAcquaintance(TalkClientContact clientContact) throws SQLException, NoClientIdInPresenceException {
-        if (clientContact.isClient() && clientContact.isNearby() && (
-                clientContact.getClientRelationship() == null
+    private void keepAcquaintance(TalkClientContact clientContact) throws SQLException, NoClientIdInPresenceException {
+        if (clientContact.isClient()  && clientContact.isInEnvironment() && (clientContact.getClientRelationship() == null
                         || clientContact.getClientRelationship().isNone()
                         || clientContact.getClientRelationship().isInvited()
                         || clientContact.getClientRelationship().invitedMe())) {
+            if (clientContact.isNearby()) {
+                clientContact.getClientPresence().setAcquaintanceType(TalkEnvironment.TYPE_NEARBY);
+            } else if (clientContact.isWorldwide()) {
+                clientContact.getClientPresence().setAcquaintanceType(TalkEnvironment.TYPE_WORLDWIDE);
+            }
             clientContact.getClientPresence().setKept(true);
-            clientContact.getClientPresence().setNearbyAcquaintance(true);
             mDatabase.savePresence(clientContact.getClientPresence());
             mDatabase.saveContact(clientContact);
         }
@@ -2793,7 +2803,7 @@ public class XoClient implements JsonRpcConnection.Listener, TransferListener {
                 clientContact.getClientPresence().setKept(false);
             } else if (friendshipCancelled(newRelationship, oldRelationShip)) {
                 clientContact.getClientPresence().setKept(true);
-                clientContact.getClientPresence().setNearbyAcquaintance(false);
+                clientContact.getClientPresence().setAcquaintanceType(TalkPresence.TYPE_ACQUAINTANCE_NONE);
             }
             clientContact.updateRelationship(newRelationship);
             mDatabase.saveRelationship(clientContact.getClientRelationship());
@@ -2865,6 +2875,8 @@ public class XoClient implements JsonRpcConnection.Listener, TransferListener {
         // quietly destroy nearby group
         if (groupPresence.isTypeNearby() && !groupPresence.exists()) {
             destroyNearbyGroup(groupContact);
+        } else if (groupPresence.isTypeWorldwide() && !groupPresence.exists()) {
+            destroyWorldwideGroup(groupContact);
         }
 
         LOG.info("updateGroupPresence(" + groupPresence.getGroupId() + ") - saved");
@@ -2905,6 +2917,40 @@ public class XoClient implements JsonRpcConnection.Listener, TransferListener {
             }
         } catch (SQLException e) {
             LOG.error("Error while destroying nearby group " + groupContact.getGroupId());
+        }
+    }
+
+    private void destroyWorldwideGroup(TalkClientContact groupContact) {
+        LOG.debug("destroying worldwide group with id " + groupContact.getGroupId());
+
+        try {
+            // reset group state
+            TalkGroupPresence groupPresence = groupContact.getGroupPresence();
+            if (groupPresence == null) {
+                LOG.error("Can not destroy worldwide group since groupPresence is null");
+                return;
+            }
+            groupPresence.setKept(true);
+
+            // save group
+            mDatabase.saveContact(groupContact);
+            mDatabase.saveGroupPresence(groupPresence);
+
+            // set member states to none
+            List<TalkGroupMembership> memberships = mDatabase.findMembershipsInGroup(groupContact.getGroupId());
+            for (TalkGroupMembership membership : memberships) {
+                membership.setState(TalkGroupMembership.STATE_NONE);
+                mDatabase.saveGroupMembership(membership);
+            }
+
+            // set worldwide state to false
+            List<TalkClientContact> contacts = mDatabase.findContactsInGroup(groupContact.getGroupId());
+            for (TalkClientContact contact : contacts) {
+                contact.setWorldwide(false);
+                mDatabase.saveContact(contact);
+            }
+        } catch (SQLException e) {
+            LOG.error("Error while destroying worldwide group " + groupContact.getGroupId());
         }
     }
 
@@ -3002,14 +3048,19 @@ public class XoClient implements JsonRpcConnection.Listener, TransferListener {
                     TalkGroupPresence groupPresence = groupContact.getGroupPresence();
                     if (groupPresence != null && groupPresence.isTypeNearby()) {
                         destroyNearbyGroup(groupContact);
+                    } else if (groupPresence != null && groupPresence.isTypeWorldwide()) {
+                        destroyWorldwideGroup(groupContact);
                     }
                 }
             }
             // if this concerns the membership of someone else
             if (clientContact.isClient()) {
-                /* Mark as nearby contact and save to database. */
+                /* Mark as nearby or worldwide contact and save to database. */
                 boolean isJoinedInNearbyGroup = groupContact.getGroupPresence() != null && groupContact.getGroupPresence().isTypeNearby() && newMembership.isJoined();
                 clientContact.setNearby(isJoinedInNearbyGroup);
+
+                boolean isJoinedInWorldwideGroup = groupContact.isWorldwideGroup() && newMembership.isJoined();
+                clientContact.setWorldwide(isJoinedInWorldwideGroup);
 
                 mDatabase.saveContact(clientContact);
             }
